carefully kill old brig
